+++ components/Shell.tsx
@@ -1,13 +1,13 @@
 import { SelectorIcon } from "@heroicons/react/outline";
 import {
   CalendarIcon,
+  ArrowLeftIcon,
   ClockIcon,
   CogIcon,
   ExternalLinkIcon,
   LinkIcon,
   LogoutIcon,
   PuzzleIcon,
-  UserGroupIcon,
 } from "@heroicons/react/solid";
 import { signOut, useSession } from "next-auth/client";
 import Link from "next/link";
@@ -36,6 +36,7 @@ import Dropdown, {
 
 import { useViewerI18n } from "./I18nLanguageHandler";
 import Logo from "./Logo";
+import Button from "./ui/Button";
 
 function useMeQuery() {
   const meQuery = trpc.useQuery(["viewer.me"], {
@@ -118,6 +119,8 @@ export default function Shell(props: {
   subtitle?: ReactNode;
   children: ReactNode;
   CTA?: ReactNode;
+  HeadingLeftIcon?: ReactNode;
+  showBackButton?: boolean;
 }) {
   const { t } = useLocale();
   const router = useRouter();
@@ -144,12 +147,6 @@ export default function Shell(props: {
       href: "/availability",
       icon: CalendarIcon,
       current: router.asPath.startsWith("/availability"),
-    },
-    {
-      name: t("teams"),
-      href: "/teams",
-      icon: UserGroupIcon,
-      current: router.asPath.startsWith("/teams"),
     },
     {
       name: t("integrations"),
@@ -261,7 +258,15 @@ export default function Shell(props: {
               </div>
             </nav>
             <div className={classNames(props.centered && "md:max-w-5xl mx-auto", "py-8")}>
+              {props.showBackButton && (
+                <div className="mx-8 mb-8">
+                  <Button onClick={() => router.back()} StartIcon={ArrowLeftIcon} color="secondary">
+                    Back
+                  </Button>
+                </div>
+              )}
               <div className="block sm:flex justify-between px-4 sm:px-6 md:px-8 min-h-[80px]">
+                {props.HeadingLeftIcon && <div className="mr-4">{props.HeadingLeftIcon}</div>}
                 <div className="w-full mb-10">
                   <h1 className="mb-1 text-xl font-bold tracking-wide text-gray-900 font-cal">
                     {props.heading}
