+++ components/team/TeamListItem.tsx
@@ -8,17 +8,21 @@ import {
 import Link from "next/link";
 import { useState } from "react";
 
+import classNames from "@lib/classNames";
 import { useLocale } from "@lib/hooks/useLocale";
 import showToast from "@lib/notification";
-import { Team } from "@lib/team";
+import { Team } from "@lib/types/team";
 
 import { Dialog, DialogTrigger } from "@components/Dialog";
 import { Tooltip } from "@components/Tooltip";
 import ConfirmationDialogContent from "@components/dialog/ConfirmationDialogContent";
 import Avatar from "@components/ui/Avatar";
 import Button from "@components/ui/Button";
-
-import Dropdown, { DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from "../ui/Dropdown";
+import Dropdown, {
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+} from "@components/ui/Dropdown";
 
 export default function TeamListItem(props: {
   onChange: () => void;
@@ -48,49 +52,57 @@ export default function TeamListItem(props: {
   return (
     team && (
       <li className="divide-y">
-        <div className="flex justify-between my-4">
-          <div className="flex">
-            <Avatar
-              size={9}
-              imageSrc={
-                props.team.logo
-                  ? props.team.logo
-                  : "https://eu.ui-avatars.com/api/?background=fff&color=039be5&name=" +
-                    encodeURIComponent(props.team.name || "")
-              }
-              alt="Team Logo"
-              className="rounded-full w-9 h-9"
-            />
-            <div className="inline-block ml-3">
-              <span className="text-sm font-bold text-neutral-700">{props.team.name}</span>
-              <span className="block -mt-1 text-xs text-gray-400">
-                {process.env.NEXT_PUBLIC_APP_URL}/team/{props.team.slug}
-              </span>
-            </div>
-          </div>
+        <div
+          className={classNames(
+            "flex justify-between py-5 px-5 items-center",
+            props.team.role !== "INVITEE" && "group hover:bg-neutral-50"
+          )}>
+          <Link href={"/settings/teams/" + props.team.id}>
+            <a className="flex-grow text-sm truncate cursor-pointer" title={`${props.team.name}`}>
+              <div className="flex">
+                <Avatar
+                  size={9}
+                  imageSrc={
+                    props.team.logo
+                      ? props.team.logo
+                      : "https://eu.ui-avatars.com/api/?background=fff&color=f9f9f9&bold=true&background=000000&name=" +
+                        encodeURIComponent(props.team.name || "")
+                  }
+                  alt="Team Logo"
+                  className="rounded-full w-9 h-9"
+                />
+                <div className="inline-block ml-3">
+                  <span className="text-sm font-bold text-neutral-700">{props.team.name}</span>
+                  <span className="block text-xs text-gray-400">
+                    {process.env.NEXT_PUBLIC_APP_URL}/team/{props.team.slug}
+                  </span>
+                </div>
+              </div>
+            </a>
+          </Link>
           {props.team.role === "INVITEE" && (
-            <div>
+            <>
               <Button type="button" color="secondary" onClick={declineInvite}>
                 {t("reject")}
               </Button>
-              <Button type="button" color="primary" className="ml-1" onClick={acceptInvite}>
+              <Button type="button" color="primary" className="ml-2" onClick={acceptInvite}>
                 {t("accept")}
               </Button>
-            </div>
+            </>
           )}
-          {props.team.role === "MEMBER" && (
-            <div>
+          <div className="flex space-x-1">
+            {props.team.role === "OWNER" && (
+              <span className="self-center px-3 py-1 mr-3 text-xs text-gray-700 capitalize bg-gray-100 border border-gray-100 rounded-md group-hover:border-gray-200">
+                {t("owner")}
+              </span>
+            )}
+            {props.team.role === "MEMBER" && (
               <Button type="button" color="primary" onClick={declineInvite}>
                 {t("leave")}
               </Button>
-            </div>
-          )}
-          {props.team.role === "OWNER" && (
-            <div className="flex space-x-4">
-              <span className="self-center h-6 px-3 py-1 text-xs text-gray-700 capitalize rounded-md bg-gray-50">
-                {t("owner")}
-              </span>
-              <Tooltip content={t("copy_link")}>
+            )}
+            {props.team.role !== "INVITEE" && (
+              <Tooltip content={t("copy_link_team")}>
                 <Button
                   onClick={() => {
                     navigator.clipboard.writeText(
@@ -98,15 +110,18 @@ export default function TeamListItem(props: {
                     );
                     showToast(t("link_copied"), "success");
                   }}
+                  className="w-10 h-10 hover:bg-white "
                   size="icon"
                   color="minimal"
                   StartIcon={LinkIcon}
                   type="button"
                 />
               </Tooltip>
+            )}
+            {props.team.role === "OWNER" && (
               <Dropdown>
-                <DropdownMenuTrigger className="w-10 h-10 p-0 border border-transparent group text-neutral-400 hover:border-gray-200">
-                  <DotsHorizontalIcon className="w-5 h-5" />
+                <DropdownMenuTrigger className="w-10 h-10 p-0 border border-transparent group text-neutral-400 hover:border-gray-200 hover:bg-white">
+                  <DotsHorizontalIcon className="w-5 h-5 group-hover:text-gray-400" />
                 </DropdownMenuTrigger>
                 <DropdownMenuContent>
                   <DropdownMenuItem>
@@ -154,8 +169,8 @@ export default function TeamListItem(props: {
                   </DropdownMenuItem>
                 </DropdownMenuContent>
               </Dropdown>
-            </div>
-          )}
+            )}
+          </div>
         </div>
       </li>
     )
