+++ components/ui/UsernameInput.tsx
@@ -7,9 +7,11 @@ interface UsernameInputProps extends React.ComponentPropsWithRef<"input"> {
 const UsernameInput = React.forwardRef<HTMLInputElement, UsernameInputProps>((props, ref) => (
   // todo, check if username is already taken here?
   <div>
-    <label htmlFor="username" className="block text-sm font-medium text-gray-700">
-      {props.label ? props.label : "Username"}
-    </label>
+    {props.label && (
+      <label htmlFor="username" className="block text-sm font-medium text-gray-700">
+        {props.label}
+      </label>
+    )}
     <div className="flex mt-1 rounded-md shadow-sm">
       <span className="inline-flex items-center px-3 text-gray-500 border border-r-0 border-gray-300 rounded-l-sm bg-gray-50 sm:text-sm">
         {process.env.NEXT_PUBLIC_APP_URL}/{props.label && "team/"}
